feat(frontend): split scroll-area, scroll-thumb
