+++ frontend/src/ui-common/scroll-area/scroll-area.tsx
@@ -8,66 +8,17 @@ import {
   onCleanup,
   createSignal,
   untrack,
-  Accessor,
   createEffect,
   on,
 } from 'solid-js';
 
 import { Dynamic, DynamicProps } from 'solid-js/web';
-
-import * as styles from './scroll-area.css';
 import { assignInlineVars } from '@vanilla-extract/dynamic';
 
-type ScrollRect = {
-  scrollWidth: number;
-  scrollHeight: number;
-  clientWidth: number;
-  clientHeight: number;
-}
-const useElementScrollRect = (maybeElement: Accessor<Element | null>) => {
-  const [rect, setRect] = createSignal<ScrollRect>({
-    scrollWidth: 0,
-    scrollHeight: 0,
-    clientWidth: 0,
-    clientHeight: 0,
-  });
+import { recordStyleToString, useElementScrollRect } from './utils';
 
-  const resizeObserver = new ResizeObserver((entries) => {
-    const target = entries[0].target;
-
-    setRect({
-      scrollWidth: target.scrollWidth,
-      scrollHeight: target.scrollHeight,
-      clientWidth: target.clientWidth,
-      clientHeight: target.clientHeight,
-    });
-  });
-
-  createEffect(on(maybeElement, (target) => {
-    if (!target) {
-      resizeObserver.disconnect();
-      return;
-    }
-
-    resizeObserver.observe(target);
-  }));
-
-  onCleanup(() => {
-    resizeObserver.disconnect();
-  });
-
-  return rect;
-};
-
-const recordStyleToString = (...styles: (Record<string, unknown> | string)[]) => {
-  return styles.reduce((acc, style) => {
-    if (typeof style === 'string') return `${acc} ${style}`;
-
-    return Object.entries(style).reduce((acc, [key, value]) => {
-      return `${acc}${key}: ${value};`;
-    }, acc);
-  }, '');
-};
+import * as styles from './scroll-area.css';
+import { ScrollThumb } from './scroll-thumb';
 
 type ScrollAreaOnlyProps = {
   edgeSize?: number;
@@ -114,31 +65,6 @@ export const ScrollArea = <
   const scrollWidth = () => rect().scrollWidth - rect().clientWidth;
   const scrollHeight = () => rect().scrollHeight - rect().clientHeight;
 
-  const thumbSize = () => {
-    if (direction() === 'vertical') {
-      const height = rect().clientHeight - local.edgeSize * 2;
-      return height / scrollHeight() * height;
-    } else {
-      const width = rect().clientWidth - local.edgeSize * 2;
-      return width / scrollWidth() * width;
-    }
-  };
-  const thumbScrollPosition = () => {
-    const offset = local.edgeSize * 2 + thumbSize();
-
-    if (direction() === 'vertical') {
-      const defaultTop = scrollPosition() * scrollHeight();
-      const offsetTop = scrollPosition() * (rect().clientHeight - offset);
-
-      return local.edgeSize + defaultTop + offsetTop;
-    }
-
-    const defaultLeft = scrollPosition() * scrollWidth();
-    const offsetLeft = scrollPosition() * (rect().clientWidth - offset);
-
-    return local.edgeSize + defaultLeft + offsetLeft;
-  };
-
   /* lifecycle */
   const handleScroll = () => {
     let newPosition = untrack(() => scrollPosition());
@@ -183,16 +109,11 @@ export const ScrollArea = <
       }), fallbackStyle())}
     >
       {local.children}
-      <div
-        classList={{
-          [styles.scrollThumb]: true,
-          [styles.scrollThumbDirectional.vertical]: direction() === 'vertical',
-          [styles.scrollThumbDirectional.horizontal]: direction() === 'horizontal',
-        }}
-        style={assignInlineVars({
-          [styles.thumbSize]: `${thumbSize()}px`,
-          [styles.thumbPosition]: `${thumbScrollPosition()}px`,
-        })}
+      <ScrollThumb
+        parent={container()}
+        position={scrollPosition()}
+        direction={direction()}
+        edgeSize={local.edgeSize}
       />
     </Dynamic>
   );
